+++ Carbon.Cryptography/Functions/Get-CPrivateKeyPermission.ps1
@@ -90,7 +90,7 @@ function Get-CPrivateKeyPermission
 
     if ($Identity)
     {
-        if( -not (Test-CIdentity -Name $Identity) )
+        if( -not (Test-CPrincipal -Name $Identity) )
         {
             $msg = "Failed to get permissions on ""${Path}"" for ""${Identity}"" because that user/group does not " +
                    'exist.'
@@ -98,7 +98,7 @@ function Get-CPrivateKeyPermission
             return
         }
 
-        $Identity = Resolve-CIdentityName -Name $Identity
+        $Identity = Resolve-CPrincipalName -Name $Identity
     }
 
     foreach ($certificate in (Get-Item -Path $Path -Force))